Add `move` function to Robot script
Add a `move` function to the Robot script that will update the
position of the Robot's `X` and `Y` properties by increasing their value
by 1 or 0, depending on which direction it facing and if the next move
does not place it outside or off the Table. (Robot's place off the table
do not move).

== Notes:
- `rune` type is Golang's character type, however is a wrapper of the
  Unicode codepoint and can include emojis or UTF-8 characters that are
  often a bit more complex

== References:
- [The Go Programming Language Specification \- The Go Programming
  Language]
  (https://golang.org/ref/spec#Rune_literals)

- [The Go Programming Language Specification \- The Go Programming
  Language]
  (https://golang.org/ref/spec#If_statements)