Add `place` function to Robot script
Add `place` function to the Robot script that will update the position
of the Robot to a new `X`, `Y` and `Facing` location based of the string
of coordinates passed in in the correct formate; e.g. `"0,0,NORTH"`.
However, will not update the Robot if the coordinates have invalid
integer values, since Go requires the return `error` value from its
String to Int method `Atoi` that then in turn needs to be used in order
for Go compiler to not complain about unused variables.

== References:
- [go \- Unpack Golang slices on assignment? \- Stack Overflow]
  (https://stackoverflow.com/questions/19832189/unpack-golang-slices-on-assignment)

- [strings \- The Go Programming Language]
  (https://golang.org/pkg/strings/#SplitN)

- [Convert string to integer type in Go? \- Stack Overflow]
  (https://stackoverflow.com/questions/4278430/convert-string-to-integer-type-in-go)

- [strconv \- The Go Programming Language]
  (https://golang.org/pkg/strconv/#Atoi)

- [strings \- The Go Programming Language]
  (https://golang.org/pkg/strings/#Contains)