Add `Main` function to Robot package
Add a `Main` function to the Robot package and update the `main.go`
entrypoint accordingly. This Main function accepts the Operating System
(OS) arguments, which is the compiled `main.go` file and any additional
arguments after its is called, when the length of those arguments is
more than just itself it will assume the first argument is a filename
and proceed to use it as its Input source. Otherwise it will default to
standard input `os.Stdin` and prompt the User for input via the console.

== Notes:
- Add several example files for testing purposes, as well for User to
  use with the package. All are based off the Toy Robot specification

== References:
- [testing \- Fill os\.Stdin for function that reads from it \- Stack
  Overflow]
  (https://stackoverflow.com/questions/46365221/fill-os-stdin-for-function-that-reads-from-it/46365584#46365584)

- [unit testing \- How can I write a Go test that writes to stdin? \-
  Stack Overflow]
  (https://stackoverflow.com/questions/37256859/how-can-i-write-a-go-test-that-writes-to-stdin)

- [go \- How can I pass a slice as a variadic input? \- Stack Overflow]
  (https://stackoverflow.com/questions/23723955/how-can-i-pass-a-slice-as-a-variadic-input)

- [go \- Read line in golang \- Stack Overflow]
  (https://stackoverflow.com/questions/40035635/read-line-in-golang)

- [go \- golang read line by line \- Stack Overflow]
  (https://stackoverflow.com/questions/30806288/golang-read-line-by-line)